[wnsxor1993#76] move: DTO 패키지 내부 분류 작업
- DTO 패키지 하위에 모든 DTO가 모여있어서 이를 하위에 내부 분류함
