+++ BE/src/main/java/com/codesquad/issuetracker/web/dto/issue/IssueCreateResponse.java
@@ -1,4 +1,4 @@
-package com.codesquad.issuetracker.web.dto;
+package com.codesquad.issuetracker.web.dto.issue;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -6,11 +6,11 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class IssueUpdateResponse {
+public class IssueCreateResponse {
 
     private Long issueId;
 
-    public IssueUpdateResponse(Long issueId) {
+    public IssueCreateResponse(Long issueId) {
         this.issueId = issueId;
     }
 }
