+++ BE/src/main/java/com/codesquad/issuetracker/web/dto/issue/IssueDetailResponse.java
@@ -1,9 +1,10 @@
-package com.codesquad.issuetracker.web.dto;
+package com.codesquad.issuetracker.web.dto.issue;
 
 import com.codesquad.issuetracker.domain.Comment;
 import com.codesquad.issuetracker.domain.Issue;
 import com.codesquad.issuetracker.domain.IssueAssignee;
 import com.codesquad.issuetracker.domain.IssueLabel;
+import com.codesquad.issuetracker.web.dto.label.LabelDto;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
