+++ BE/src/main/java/com/codesquad/issuetracker/web/dto/issue/IssueListResponse.java
@@ -1,7 +1,8 @@
-package com.codesquad.issuetracker.web.dto;
+package com.codesquad.issuetracker.web.dto.issue;
 
 import com.codesquad.issuetracker.domain.Issue;
 import com.codesquad.issuetracker.domain.IssueLabel;
+import com.codesquad.issuetracker.web.dto.label.LabelDto;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
