+++ BE/src/main/java/com/codesquad/issuetracker/web/dto/issue/IssueUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.codesquad.issuetracker.web.dto;
+package com.codesquad.issuetracker.web.dto.issue;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -12,7 +12,7 @@ import java.util.List;
 @Getter
 @ToString
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class IssueCreateRequest {
+public class IssueUpdateRequest {
 
     @NotBlank
     private String title;
