+++ BE/src/main/java/com/codesquad/issuetracker/web/dto/issue/IssueUpdateResponse.java
@@ -1,4 +1,4 @@
-package com.codesquad.issuetracker.web.dto;
+package com.codesquad.issuetracker.web.dto.issue;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -6,11 +6,11 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class IssueCreateResponse {
+public class IssueUpdateResponse {
 
     private Long issueId;
 
-    public IssueCreateResponse(Long issueId) {
+    public IssueUpdateResponse(Long issueId) {
         this.issueId = issueId;
     }
 }
