+++ BE/src/main/java/com/codesquad/issuetracker/web/dto/issue/MultipleIssueStateChangeResponse.java
@@ -1,4 +1,4 @@
-package com.codesquad.issuetracker.web.dto;
+package com.codesquad.issuetracker.web.dto.issue;
 
 import com.codesquad.issuetracker.domain.Issue;
 import lombok.AccessLevel;
